refactor proposal memo reporting routes
